+++ h2integrate/converters/hydrogen/pem_model/PEM_H2_LT_electrolyzer_Clusters.py
@@ -27,7 +27,6 @@ import scipy
 import pandas as pd
 import rainflow
 from scipy import interpolate
-from matplotlib import pyplot as plt
 
 
 np.set_printoptions(threshold=sys.maxsize)
@@ -992,202 +991,3 @@ class PEM_H2_Clusters:
         h2_results_aggregates["Time until replacement [hours]"] = time_until_replacement
 
         return h2_results, h2_results_aggregates
-
-
-if __name__ == "__main__":
-    cluster_size_mw = 1  # MW
-    plant_life = 30  # years
-    electrolyzer_model_parameters = {
-        "eol_eff_percent_loss": 10,  # percent - for calculating EOL voltage and steady deg rate
-        "uptime_hours_until_eol": 77600,  # for calculating steady deg rate
-        "include_degradation_penalty": True,
-        "turndown_ratio": 0.1,
-        "curve_coeff": [
-            4.0519644766515644e-08,
-            -0.00026186723338675105,
-            3.8985774154190334,
-            7.615382921418666,
-            -20.075110413404484,
-            1.0,
-        ],
-    }
-    # Create PEM and initialize parameters
-    pem = PEM_H2_Clusters(cluster_size_mw, plant_life, **electrolyzer_model_parameters)
-    # ----- Run off-grid case -----
-    # Make a mock input power signal
-    hours_in_year = 8760
-    power_signal_kW_ramp_up = np.arange(0, 1000 * cluster_size_mw, 20)
-    power_signal_kW_ramp_down = np.flip(power_signal_kW_ramp_up)
-    power_signal_kW_ramp_updown = np.concatenate(
-        (power_signal_kW_ramp_down, power_signal_kW_ramp_up)
-    )
-    n_repeats = int(np.ceil(hours_in_year / len(power_signal_kW_ramp_updown)))
-    power_signal_kW = np.tile(power_signal_kW_ramp_updown, n_repeats)
-    # Run the electrolyzer and get outputs
-    h2_results, h2_results_aggregates = pem.run(power_signal_kW)
-
-    # ----- Run grid-connected case -----
-    h2_kg_hr_system_required = 12  # kg-H2/hr
-    target_h2_per_year = h2_kg_hr_system_required * hours_in_year
-    power_required_kW, stack_current_signal = pem.grid_connected_func(h2_kg_hr_system_required)
-    h2_results_grid, h2_results_aggregates_grid = pem.run_grid_connected_workaround(
-        power_required_kW, stack_current_signal
-    )
-
-    # Refurbishment Schedule (RS) [MW/year] and Year of Replacement (YOR)
-    offgrid_RS = np.array(
-        list(
-            h2_results_aggregates["Performance By Year"][
-                "Refurbishment Schedule [MW replaced/year]"
-            ].values()
-        )
-    )
-    offgrid_YOR = np.argwhere(offgrid_RS > 0)[:, 0]
-    grid_RS = np.array(
-        list(
-            h2_results_aggregates_grid["Performance By Year"][
-                "Refurbishment Schedule [MW replaced/year]"
-            ].values()
-        )
-    )
-    grid_YOR = np.argwhere(grid_RS > 0)[:, 0]
-    # ----- Plot and compare results -----
-    year_of_operation = list(
-        h2_results_aggregates["Performance By Year"]["Capacity Factor [-]"].keys()
-    )
-    fig, ax = plt.subplots(nrows=3, ncols=1, sharex=True, figsize=[6.4, 7.2])
-
-    # Annual Hydrogen Produced (AH2) [metric tons H2/year]
-    offgrid_AH2 = (
-        np.array(
-            list(
-                h2_results_aggregates["Performance By Year"][
-                    "Annual H2 Production [kg/year]"
-                ].values()
-            )
-        )
-        / 1000
-    )
-    grid_AH2 = (
-        np.array(
-            list(
-                h2_results_aggregates_grid["Performance By Year"][
-                    "Annual H2 Production [kg/year]"
-                ].values()
-            )
-        )
-        / 1000
-    )
-
-    ax[0].plot(year_of_operation, offgrid_AH2, color="b", ls="-", lw=2, label="off-grid")
-    ax[0].plot(year_of_operation, grid_AH2, color="r", ls="--", lw=2, label="grid-connected")
-    ax[0].vlines(
-        offgrid_YOR,
-        np.zeros(len(offgrid_YOR)),
-        offgrid_AH2[offgrid_YOR],
-        colors="skyblue",
-        ls="-",
-        lw=0.5,
-    )
-    ax[0].vlines(
-        grid_YOR,
-        np.zeros(len(grid_YOR)),
-        grid_AH2[grid_YOR],
-        colors="tomato",
-        ls="--",
-        lw=0.5,
-    )
-    ax[0].set_ylabel("Annual Hydrogen Produced\n[metric tons/year]")
-    ax[0].legend(loc="upper right")
-    ax[0].set_ylim((0.8 * np.min([offgrid_AH2, grid_AH2]), 1.2 * np.max([offgrid_AH2, grid_AH2])))
-
-    # Annual Energy Used (AEU) [MWh/year]
-    offgrid_AEU = (
-        np.array(
-            list(
-                h2_results_aggregates["Performance By Year"][
-                    "Annual Energy Used [kWh/year]"
-                ].values()
-            )
-        )
-        / 1000
-    )
-    grid_AEU = (
-        np.array(
-            list(
-                h2_results_aggregates_grid["Performance By Year"][
-                    "Annual Energy Used [kWh/year]"
-                ].values()
-            )
-        )
-        / 1000
-    )
-    ax[1].plot(year_of_operation, offgrid_AEU, color="b", ls="-", lw=2, label="_off-grid")
-    ax[1].plot(year_of_operation, grid_AEU, color="r", ls="--", lw=2, label="_grid-connected")
-    ax[1].vlines(
-        offgrid_YOR,
-        np.zeros(len(offgrid_YOR)),
-        offgrid_AEU[offgrid_YOR],
-        colors="skyblue",
-        ls="-",
-        lw=0.5,
-        label="off-grid stack replacement",
-    )
-    ax[1].vlines(
-        grid_YOR,
-        np.zeros(len(grid_YOR)),
-        grid_AEU[grid_YOR],
-        colors="tomato",
-        ls="--",
-        lw=0.5,
-        label="grid-connected stack replacement",
-    )
-    ax[1].set_ylabel("Annual Energy Used\n[MWh/year]")
-    ax[1].legend(loc="upper right")
-    ax[1].set_ylim((0.8 * np.min([offgrid_AEU, grid_AEU]), 1.2 * np.max([offgrid_AEU, grid_AEU])))
-
-    # Annual Average Conversion Efficiency (AACE) [kWh/kg]
-    offgrid_AACE = np.array(
-        list(
-            h2_results_aggregates["Performance By Year"][
-                "Annual Average Efficiency [kWh/kg]"
-            ].values()
-        )
-    )
-    grid_AACE = np.array(
-        list(
-            h2_results_aggregates_grid["Performance By Year"][
-                "Annual Average Efficiency [kWh/kg]"
-            ].values()
-        )
-    )
-    ax[2].plot(year_of_operation, offgrid_AACE, color="b", ls="-", lw=2, label="_off-grid")
-    ax[2].plot(year_of_operation, grid_AACE, color="r", ls="--", lw=2, label="_grid-connected")
-    ax[2].vlines(
-        offgrid_YOR,
-        np.zeros(len(offgrid_YOR)),
-        offgrid_AACE[offgrid_YOR],
-        colors="skyblue",
-        ls="-",
-        lw=0.5,
-        label="off-grid stack replacement",
-    )
-    ax[2].vlines(
-        grid_YOR,
-        np.zeros(len(grid_YOR)),
-        grid_AACE[grid_YOR],
-        colors="tomato",
-        ls="--",
-        lw=0.5,
-        label="grid-connected stack replacement",
-    )
-    ax[2].set_ylabel("Annual Average Efficiency\n[kWh/kg]")
-    ax[2].set_xlabel("Year of Operation")
-    ax[2].set_xlim((0, plant_life))
-    ax[2].set_ylim(
-        (
-            0.95 * np.min([offgrid_AACE, grid_AACE]),
-            1.05 * np.max([offgrid_AACE, grid_AACE]),
-        )
-    )
-    fig.tight_layout()
