--- h2integrate/simulation/technologies/hydrogen/electrolysis/__init__.py
@@ -1,13 +1,0 @@
-import h2integrate.simulation.technologies.hydrogen.electrolysis.PEM_H2_LT_electrolyzer_Clusters
-import h2integrate.simulation.technologies.hydrogen.electrolysis.PEM_tools
-from h2integrate.simulation.technologies.hydrogen.electrolysis.PEM_costs_Singlitico_model import (
-    PEMCostsSingliticoModel,
-)
-
-# FIXME: duplicative imports
-from h2integrate.simulation.technologies.hydrogen.electrolysis.PEM_electrolyzer_IVcurve import (
-    PEM_electrolyzer_LT,
-)
-from h2integrate.simulation.technologies.hydrogen.electrolysis.PEM_H2_LT_electrolyzer import (
-    PEM_electrolyzer_LT,  # FIXME: duplicative import, delete whole comment when fixed # noqa: F811
-)
